refactor: create 'content' package
Moving Content.java, ContentComparator.java and ContentType.java to 'content' package.
