+++ src/main/java/com/github/savitoh/ApiResult.java
@@ -1,5 +1,6 @@
 package com.github.savitoh;
 
+import com.github.savitoh.content.Content;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
