+++ src/main/java/com/github/savitoh/JsonParser.java
@@ -1,6 +1,7 @@
 package com.github.savitoh;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
+import com.github.savitoh.content.Content;
 import java.util.List;
 
 public interface JsonParser {
