+++ src/main/java/com/github/savitoh/content/ContentComparator.java
@@ -1,4 +1,4 @@
-package com.github.savitoh;
+package com.github.savitoh.content;
 
 import java.util.Comparator;
