+++ src/main/java/com/github/savitoh/content/ContentType.java
@@ -1,4 +1,4 @@
-package com.github.savitoh;
+package com.github.savitoh.content;
 
 public enum ContentType {
   MOVIE,
