+++ src/main/java/com/github/savitoh/imdb/ImdbMoviesJsonParser.java
@@ -4,8 +4,8 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.github.savitoh.Content;
 import com.github.savitoh.JsonParser;
+import com.github.savitoh.content.Content;
 import java.util.List;
 import java.util.Objects;
 import java.util.regex.Matcher;
