+++ src/main/java/com/github/savitoh/marvel/Series.java
@@ -1,7 +1,7 @@
 package com.github.savitoh.marvel;
 
-import com.github.savitoh.Content;
-import com.github.savitoh.ContentType;
+import com.github.savitoh.content.Content;
+import com.github.savitoh.content.ContentType;
 
 public record Series(String title, String urlImage, String rating, String year)
     implements Content, Comparable<Content> {
