+++ src/main/java/com/github/savitoh/marvel/SeriesJsonParser.java
@@ -3,8 +3,8 @@ package com.github.savitoh.marvel;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.github.savitoh.Content;
 import com.github.savitoh.JsonParser;
+import com.github.savitoh.content.Content;
 import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
